fix: 메인 소켓 위치 옮김
